Share the emoji properties enum between the library and the generator
