+++ tools/Meziantou.Framework.Unicode.Generator/Program.cs
@@ -1313,18 +1313,6 @@ internal sealed record Entry(Rune Source, string Target);
 
 internal sealed record PendingRange(int Start, UnicodeDataEntry Entry);
 
-[Flags]
-internal enum EmojiProperties : byte
-{
-    None = 0,
-    Emoji = 1 << 0,
-    EmojiPresentation = 1 << 1,
-    EmojiModifier = 1 << 2,
-    EmojiModifierBase = 1 << 3,
-    EmojiComponent = 1 << 4,
-    ExtendedPictographic = 1 << 5,
-}
-
 internal sealed record UnicodeDataEntry(
     Rune Rune,
     string Name,
